backend_2/requirements.txt
pandas
fastapi
uvicorn

google-auth
grpcio
fsspec

python-jose[cryptography]
passlib[bcrypt]
python-multipart>=0.0.9
psycopg2-binary
python-dotenv
requests
pydantic-settings
PyPDF2
python-docx
langchain
langchain-community
langchain-ollama
langchain-chroma
langchain-groq
chromadb
ollama
unstructured
pdf2image
pypdf
openpyxl
langgraph
langchain-core
langgraph-checkpoint-sqlite
langgraph-checkpoint-postgres
google-generativeai
langchain-huggingface
# langchain-mcp-adapters  # Optional - causes dependency conflicts, disabled for now
sentence-transformers
duckduckgo-search
wikipedia
langchain-tavily
apscheduler
langchain-google-genai
langchain-openai
huggingface-hub
Pillow
gradio_client
--extra-index-url https://download.pytorch.org/whl/cpu
